resolução exercicios

+++ CURSOEMVIDEO/ex029.py
@@ -1,5 +1,5 @@
 #Escreva um programa que leia a velocidade de um carro. Se ele ultrapassar 80km/h mostre uma mensagem dizendo que ele foi multado. A multa vai custa R$7,00 por cada km acima do limite.
-v = int(input('Entre com a velocidade em Km/h: '))
+v = float(input('Entre com a velocidade em Km/h: '))
 m = v - 80
 if v >= 81:
     print('xxxxxxxxxx LIMITE DE VELOCIDADE 80 KM/H xxxxxxxxxx')
